refactor: extract AccountContactsApi from accounts.rs (#444)
