+++ api/src/openapi.rs
@@ -1,4 +1,5 @@
 pub mod accounts;
+pub mod accounts_contacts;
 pub mod accounts_keys;
 pub mod admin;
 pub mod allowlist;
@@ -33,6 +34,7 @@ pub mod webhooks;
 pub mod webhooks_disputes;
 
 pub use accounts::AccountsApi;
+pub use accounts_contacts::AccountContactsApi;
 pub use accounts_keys::AccountKeysApi;
 pub use admin::AdminApi;
 pub use allowlist::AllowlistApi;
@@ -102,6 +104,7 @@ pub fn create_combined_api() -> impl OpenApi {
             RecoveryApi,
             EmailVerificationApi,
             AccountKeysApi,
+            AccountContactsApi,
         ),
     )
 }
